+++ ohara-manager/scripts/createServices.js
@@ -17,7 +17,7 @@
 /* eslint-disable no-process-exit, no-console */
 
 const yargs = require('yargs');
-const axios = require('axios');
+const api = require('../utils/apiHandler');
 const fs = require('fs');
 
 const {
@@ -29,8 +29,8 @@ const {
   nodePass,
 } = yargs.argv;
 
-const zkName = 'zk' + randomName();
-const bkName = 'bk' + randomName();
+const zkName = 'zk' + api.randomName();
+const bkName = 'bk' + api.randomName();
 
 debug('configurator: ', configurator);
 debug('port: ', port);
@@ -44,12 +44,19 @@ function debug(...message) {
 }
 
 const createServices = async () => {
-  await createNode();
-  await wait('nodes', nodeHost);
-  await createZk();
-  await wait('zookeepers', zkName);
-  await createBk();
-  await wait('brokers', bkName);
+  await api.createNode(
+    configurator,
+    port,
+    nodeHost,
+    nodePort,
+    nodeUser,
+    nodePass,
+  );
+  await api.waitCreate(configurator, port, 'nodes', nodeHost);
+  await api.createZk(configurator, port, zkName, nodeHost);
+  await api.waitCreate(configurator, port, 'zookeepers', zkName);
+  await api.createBk(configurator, port, zkName, bkName, nodeHost);
+  await api.waitCreate(configurator, port, 'brokers', bkName);
   await fs.writeFile(
     'scripts/env/service.json',
     `{"zk":"${zkName}","bk":"${bkName}"}`,
@@ -60,70 +67,3 @@ const createServices = async () => {
 };
 
 createServices();
-
-async function createNode() {
-  await axios.post(`http://${configurator}:${port}/v0/nodes`, {
-    name: nodeHost,
-    port: nodePort,
-    user: nodeUser,
-    password: nodePass,
-  });
-}
-
-async function createZk() {
-  await axios.post(`http://${configurator}:${port}/v0/zookeepers`, {
-    clientPort: randomPort(),
-    electionPort: randomPort(),
-    peerPort: randomPort(),
-    name: zkName,
-    nodeNames: [nodeHost],
-  });
-}
-
-async function createBk() {
-  await axios.post(`http://${configurator}:${port}/v0/brokers`, {
-    clientPort: randomPort(),
-    exporterPort: randomPort(),
-    jmxPort: randomPort(),
-    zookeeperClusterName: zkName,
-    name: bkName,
-    nodeNames: [nodeHost],
-  });
-}
-
-async function wait(api, name) {
-  const res = await axios.get(`http://${configurator}:${port}/v0/` + api);
-  var result = res.data.some(e => e.name == name);
-
-  if (!result) {
-    sleep(1000);
-    await wait(api, name);
-  }
-
-  return;
-}
-
-function randomPort() {
-  var min = 5000;
-  var max = 65535;
-  return Math.floor(Math.random() * (max - min + 1)) + min;
-}
-
-function randomName() {
-  var text = '';
-  var possible = 'abcdefghijklmnopqrstuvwxyz0123456789';
-
-  for (var i = 0; i < 5; i++)
-    text += possible.charAt(Math.floor(Math.random() * possible.length));
-
-  return text;
-}
-
-function sleep(milliseconds) {
-  var start = new Date().getTime();
-  for (var i = 0; i < 1e7; i++) {
-    if (new Date().getTime() - start > milliseconds) {
-      break;
-    }
-  }
-}
